% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.r
\name{death}
\alias{death}
\title{title: Model definitions}
\usage{
death(rho, parms)
colonization(rho, parms)
}
\description{
title: Model definitions
}
\details{
defining chances of death and colonization per location, depending on the global and local vegetation cover.
}

