% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equilibria.R
\name{get_equilibria}
\alias{get_equilibria}
\title{Title}
\usage{
get_equilibria(odesys, y, parms, t_max = 1000, method = "ode45")
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{

get_equilibria(livestock$pair, ini_rho(rho_1 = 0.9999), livestock$parms)
get_equilibria(livestock$meanfield, ini_rho(rho_1 = 0.9999), livestock$parms)
}

