% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.r
\name{highlight}
\alias{highlight}
\title{Title}
\usage{
highlight(x, colrange = c("black", "red2"), steps = NULL, range = "auto")
}
\arguments{
\item{x}{}

\item{colrange}{}

\item{steps}{}

\item{range}{}
}
\description{
Title
}

