% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.r
\name{mortality}
\alias{mortality}
\title{calculate plant mortality at a given cover and local cover}
\usage{
mortality(rho_1, q_11 = 1, parms = livestock$defparms)
}
\arguments{
\item{set}{}
}
\description{
calculate plant mortality at a given cover and local cover
}

