% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_base.r
\name{plot_base}
\alias{plot_base}
\title{Base plot of growth and mortality over vegetation cover}
\usage{
plot_base(...)
}
\arguments{
\item{...}{}
}
\description{
Base plot of growth and mortality over vegetation cover
}

