% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meanfield.r
\name{plot_meanfield}
\alias{plot_meanfield}
\title{Draw meanfield model attractor}
\usage{
plot_meanfield(model, func = model$meanfield, parms = model$parms,
  times = c(0, 1000), method = "ode45", rho_x = seq(0, 1, length = 100),
  col = c("#000000", "#009933"), new = FALSE)
}
\arguments{
\item{colors}{}
}
\description{
Draw meanfield model attractor
}
\examples{

plot_meanfield(livestock)

}

