% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairapproximation.r
\name{plot_pairapproximation}
\alias{plot_pairapproximation}
\title{Draw meanfield model attractor}
\usage{
plot_pairapproximation(model, parms = model$parms, side = "rho",
  rho_1_ini = seq(0, 1, length = 11), rho_11_ini = seq(0, 1, length = 11),
  times = c(0, 1000), method = "ode45", rho = seq(0, 1, length = 100),
  colors = c("#000000", "#009933"), new = TRUE, ...)
}
\arguments{
\item{colors}{}
}
\description{
Draw meanfield model attractor
}
\examples{

p <- set_parms(livestock$parms, set = list(b = 0.1, f = 0.5, p = 0.9, L = 1.5))
par(mfrow = c(1,2))
plot_pairapproximation(livestock, parms = p) -> out
plot_pairapproximation(out, parms = p, side = "plain")
}

