% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_11.r
\name{q_01}
\alias{q_01}
\title{Transfer cover object into average local cover of empty cells, q_01.}
\usage{
q_01(rho)
}
\arguments{
\item{rho}{}
}
\description{
Transfer cover object into average local cover of empty cells, q_01.
}
\examples{

rho <- ini_rho(0.5,0.45)
q_01(rho)

}

