% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_11.r
\name{q_11}
\alias{q_11}
\title{Transfer cover object into average local cover of vegetated cells, q_11.}
\usage{
q_11(rho)
}
\arguments{
\item{rho}{}
}
\description{
Transfer cover object into average local cover of vegetated cells, q_11.
}
\examples{

rho <- ini_rho(0.5,0.45)
q_11(rho)
}

