% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ode.R
\name{run_ode}
\alias{run_ode}
\title{Title}
\usage{
run_ode(y, func, times = 1.05^seq(0, 100, 1), parms, ...)
}
\arguments{
\item{parms}{}
}
\description{
Title
}
\examples{
p <- set_parms(livestock$parms, set = list(b = 0.2, f = 0.9, p = 0.8))
out <- run_ode(ini_rho(0.0001), livestock$pair, parms = p)
plot(out[,1:2], type = "l", ylim = c(0,1))
}

