% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parms.r
\name{set_parms}
\alias{set_parms}
\title{Title}
\usage{
set_parms(parms, set = list(NA))
}
\arguments{
\item{set}{}
}
\description{
Title
}
\examples{

p <- set_parms(livestock$defparms, set = list(b = 0.2, f = 0.9))

}

