% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ode.R, R/sim.r
\name{sim}
\alias{sim}
\title{Title}
\usage{
sim(input, model, times = exp(seq(0, 4, length = 100)) - 1,
  parms = model$defparms)

sim(input, model, times = exp(seq(0, 4, length = 100)) - 1,
  parms = model$defparms)
}
\arguments{
\item{parms}{}

\item{parms}{}
}
\description{
Title

Title
}
\examples{

sim(ini_rho(0.2,0.2), livestock)
}

