% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bifurcations.r
\name{sim_bifurcations}
\alias{sim_bifurcations}
\title{Title}
\usage{
sim_bifurcations(model, parms = model$defparms, over = "b", xrange = c(0,
  1), res = 31, ini = c(0.9, 1e-04), t_max = 150, func = model$pair,
  method = "ode45")
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{

library(foreach)
library(doSNOW)

workerlist <- c(rep("localhost", times = 3))
cl <- makeSOCKcluster(workerlist)
registerDoSNOW(cl)

p <- set_parms(livestock$defparms, set = list(b = 0.9, c = 0.2, f = 0, p = 0, alpha = 0.2))
sim_bifurcations(livestock, over = "b", res = 31, parms = p)

stopCluster(cl)

}

