% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trajectories.r
\name{sim_trajectories}
\alias{sim_trajectories}
\title{Title}
\usage{
sim_trajectories(model, parms = model$parms, rho_1_ini = seq(0, 1, length =
  11), times = c(0, 1000), func = model$pair, method = "ode45")
}
\arguments{
\item{method}{}
}
\description{
Title
}
\examples{

library(foreach)
p <- set_parms(livestock$parms, set = list(b = 0.2,f = 0.8))
sim_trajectories(livestock, parms = p)


}

