% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_rho.r
\name{ini_rho}
\alias{ini_rho}
\title{Title}
\usage{
ini_rho(rho_1, rho_11 = NULL, cc = NULL)
}
\arguments{
\item{rho_1}{A single numerical value between 0 and 1. Total vegetation cover, i.e. the probability of a cell to be vegetated.}

\item{rho_11}{A single numerical value between 0 and 1. Average local vegetation cover of vegetated cells, i.e.  the conditional probability of finding vegetated cells in the 4-cell neighborhood given that the focal cell is vegetated. It defaults to NULL which leads to an assumption of neutral clustering, i.e. \code{rho_11 == rho_1}.}

\item{cc}{A single numerical value greater than 0. Clustering coefficient,
describing the ratio between average local cover \code{rho_11}, and global cover \code{rho_1}. A value greater 1 denotes strong clustering. It defaults to NULL which leads to an assumption of neutral clustering, i.e. \code{rho_11 == rho_1}.}
}
\value{
A named vector of
}
\description{
Title
}
\examples{

ini_rho(0.9, cc = 1.2)


}

