% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bifurcation.r
\name{plot_bifurcation}
\alias{plot_bifurcation}
\title{Title}
\usage{
plot_bifurcation(model, parms = model$parms, over = "b",
  type = "bifurcation", xrange = c(0, 1), res = 201, ini = c(0.9,
  1e-04), t_max = 150, method = "ode45", xlab = over, unstable = TRUE,
  colors = c("#000000", "#009933"), new = TRUE, ...)
}
\arguments{
\item{new}{}
}
\description{
Title
}
\examples{

library(foreach)
library(doSNOW)

workerlist <- c(rep("localhost", times = 3))
cl <- makeSOCKcluster(workerlist)
registerDoSNOW(cl)

p <- set_parms(livestock$parms, set = list(b = 0.2, f = 0.9, p = 0.3, c = 0.2))
par(mfrow = c(1,2))
plot_bifurcation(livestock, over = "L", res = 101, xrange = c(0,3), parms = p) -> out
plot_bifurcation(out, over = "L", res = 101, xrange = c(0,3), type = "plain", parms = p)

stopCluster(cl)


}

