% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairapproximation3D.R
\name{plot_pairapproximation3D}
\alias{plot_pairapproximation3D}
\title{Draw meanfield model attractor}
\usage{
plot_pairapproximation3D(model, parms = model$defparms, rho_1_ini = seq(0,
  1, length = 21), rho_11_ini = seq(0, 1, length = 11), times = c(0, 1000),
  method = "ode45", colors = c("#000000", "#009933"))
}
\arguments{
\item{colors}{}
}
\description{
Draw meanfield model attractor
}
\examples{

p <- set_parms(livestock$defparms, set = list(b = 0.2, c = 0.2, f = 0.9, p = 0.99))
plot_pairapproximation3D(livestock, parms = p)
}

